Add /plugin slash command with autocomplete
Refactor plugin command to extract shared utilities into pluginUtils.tsx,
eliminating code duplication between text command and slash command.
